@@ -8,8 +8,9 @@
 
 namespace Spoonity\Controller\Abstraction;
 
-
+use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -18,6 +19,30 @@ use Symfony\Component\HttpFoundation\Request;
  */
 abstract class BaseController extends AbstractController
 {
+    /** @var ManagerRegistry  */
+    private ManagerRegistry $manager;
+
+    /** @var EventDispatcherInterface  */
+    protected EventDispatcherInterface $dispatcher;
+
+    /**
+     * @param ManagerRegistry $manager
+     * @param EventDispatcherInterface $dispatcher
+     */
+    public function __construct(ManagerRegistry $manager, EventDispatcherInterface $dispatcher)
+    {
+        $this->manager = $manager;
+        $this->dispatcher = $dispatcher;
+    }
+
+    /**
+     * @return ManagerRegistry
+     */
+    public function getManager(): ManagerRegistry
+    {
+        return $this->manager;
+    }
+
     /**
      * @param Request $request
      * @return int
@@ -63,5 +88,14 @@ abstract class BaseController extends AbstractController
         }
 
         return $order;
+    }
+
+    /**
+     * @param Request $request
+     * @return array
+     */
+    public function getContent(Request $request): array
+    {
+        return json_decode(($request->getContent()), true);
     }
 }
