@@ -18,7 +18,7 @@ interface Meta
     /**
      * @return int
      */
-    public function getId(): int;
+    public function getId(): ?int;
 
     /**
      * @return string
